requirements.txt
requests
beautifulsoup4
python3-nmap
dnspython
python-owasp-zap-v2.4

